basic working split of Sketch class
